Add messages and show messages on the timeline. Refine index.html

+++ src/main.rs
@@ -1,7 +1,6 @@
 use actix_web::middleware::Logger;
 use actix_web::{App, HttpServer};
 
-mod auth;
 mod config;
 mod db;
 mod error;
@@ -16,23 +15,27 @@ pub type Result<T, E = error::Error> = std::result::Result<T, E>;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    if let Ok((config, db, mimetypes)) = initialize() {
-        let config_clone = config.clone();
-        HttpServer::new(move || {
-            App::new()
-                .wrap(Logger::new(&config_clone.logging_template))
-                .data(pages::create_handlebars())
-                .data(db.clone())
-                .data(config_clone.clone())
-                .data(mimetypes.clone())
-                .configure(pages::config)
-        })
-        .bind((config.host.clone(), config.port))?
-        .run()
-        .await
-    } else {
-        log::info!("Shutting down...");
-        Ok(())
+    match initialize() {
+        Ok((config, db, mimetypes)) => {
+            let config_clone = config.clone();
+            HttpServer::new(move || {
+                App::new()
+                    .wrap(Logger::new(&config_clone.logging_template))
+                    .data(pages::create_handlebars())
+                    .data(db.clone())
+                    .data(config_clone.clone())
+                    .data(mimetypes.clone())
+                    .configure(pages::config)
+            })
+            .bind((config.host.clone(), config.port))?
+            .run()
+            .await
+        }
+        Err(e) => {
+            log::error!("{}", e);
+            log::info!("Shutting down...");
+            Ok(())
+        }
     }
 }
 
@@ -45,8 +48,10 @@ fn initialize() -> Result<(Config, Database, MimeTypes)> {
     env_logger::init();
 
     let mut db = Database::new(config.clone());
-    if let Err(e) = db.drop() {
-        log::warn!("Could not drop db: {}", e);
+    if config.force_recreate_db {
+        if let Err(e) = db.drop() {
+            log::warn!("Could not drop db: {}", e);
+        }
     }
     db.initialize()?;
